requirements.txt
# Core dependencies
pydantic==2.5.3
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
tenacity==8.2.3

# Data processing
pandas==2.1.4
numpy==1.26.3

# API and web
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.2.0
hypothesis==6.96.1

# Visualization and reporting
plotly==5.18.0
rich==13.7.0
tabulate==0.9.0

# Development tools
black==23.12.1
isort==5.13.2
mypy==1.8.0
flake8==7.0.0
pre-commit==3.6.0

# Logging and monitoring
structlog==24.1.0
python-json-logger==2.0.7